fix: authView bug
I'm working on a bug that where hitting cancel in the authView just leads a second authView to
pop up immediately after dismissal. I think implementing the authView at the button press rather
than the the viewDidLoad of the TabBar will fix it.

TODO:
- Finish implementing the loginSession in the HomeMapVC
- Clean up the code in VendorInfoVC that checks if they are logged in to give errorAlert rather
than loginSession, due to the fix.