feat: adding InfoVC to the vendorTabVC
working on adding buttons for saving and opening to navbar
of the vendortabVC. Also adding a the FoodTruckInfoView to the tab bar,
so a vendor can see what their profile actually looks like.

issues:
the tab bar is creating an extra "profile" tab, and i'm not sure why.
buttons still need to be wired up.
vendor persistence should be switched from CoreDdata to Firebase to ensure
easier communication/ less errors.